+++ CHANGELOG.md
@@ -69,6 +69,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - GUI for met data merger. [#266](https://github.com/ncsa/ameriflux-pipeline/issues/266)
 - Timestamp checks for pyfluxpro input sheet processing. [#267](https://github.com/ncsa/ameriflux-pipeline/issues/267)
 - Met tower variable name mapping for 2020 data. [#260](https://github.com/ncsa/ameriflux-pipeline/issues/260)
+- Check for duplicate variable names in L1 and L2. [#263](https://github.com/ncsa/ameriflux-pipeline/issues/263)
 
 ### Fixed
 - Python packaging of the codes so it can run in command prompt. [#70](https://github.com/ncsa/ameriflux-pipeline/issues/70)